# build: switch to next-rspack