# Summary

<!--
Please include a summary of the change and which issue is fixed. Please also include relevant motivation and context. List any dependencies that are required for this change.

> [!Tip]
> Close issue syntax: `Fixes #<issue number>` or `Resolves #<issue number>`, see [documentation](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword) for more details.
-->

This PR introduces [next-rspack](https://rspack.dev/guide/tech/next) to improve Next.js performance. **We should be careful since it is experimental now.**

Close https://github.com/langgenius/dify/issues/19179.

# Screenshots

| Before | After |
|--------|-------|
| <img width="546" alt="Image" src="https://github.com/user-attachments/assets/7e6940ca-b390-4f58-8e70-2c723010c99d" />    | <img width="502" alt="image" src="https://github.com/user-attachments/assets/21d2fffa-ad70-4ebb-add0-bea2bd5ddf1d" /> |

`/apps`: 47.1s -> 23.9s, 49.25% Faster
`/app/[appId]/workflow`: 20.5s -> 10.3s, 49.76% Faster
`/app/[appId]/develop`: 13.8s -> 3.4s, 75.36% Faster
`/datasets`: 15.3s -> 2.8s, 81.70% Faster

# Checklist

> [!IMPORTANT]  
> Please review the checklist below before submitting your pull request.

- [ ] This change requires a documentation update, included: [Dify Document](https://github.com/langgenius/dify-docs)
- [x] I understand that this PR may be closed in case there was no previous discussion or issues. (This doesn't apply to typos!)
- [x] I've added a test for each change that was introduced, and I tried as much as possible to make a single atomic change.
- [x] I've updated the documentation accordingly.
- [x] I ran `dev/reformat`(backend) and `cd web && npx lint-staged`(frontend) to appease the lint gods